Fix "No row was returned for one()" database error
If a user tried to access the edit/delete pages for a category or an
item which doesn't even exist in the database by modifying the params in
the url link, the page shown the error of "Now row returned for one()".
It was fixed by changing the ".one()" function to ".one_or_none()"
function and making a small check if the category/item exists in the
database.